Ramen Shop
Have the basic gameplay running.
Reworked the ingredient system and my general idea for the game.
Implemented a lot of scriptable object classes and classes to contain them. Got the ramen Opinion system more inline with what I am thinking and just played with a lot of systems to try and see what works and what doesnt. I think now Im in a good spot to start working on some more complimentary systems like currency and items.
Still alot of balancing that needs to be done, but likely wont do that for a while